requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions>=1.23.0
requests>=2.31.0
numpy>=2.3.2
pydantic>=2.11.7
scikit-learn>=1.7.1

